+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## Unreleased
 
+* Added **`cleanupOnlyLastBuild`** option
 * Package updates and cleanup
 
 ## [v0.5.1](https://github.com/gpbl/webpack-cleanup-plugin/tree/v0.5.1) (2017-02-27)